Make apollo-env a standard TS project
This commit removes all of the extraneous work being done for
compiling apollo-env's global `fetch` types. To accomplish this,
`apollo-env` was converted to a standard TS project using
references, just as the other packages are implemented within this
repo.
