--- packages/apollo-env/src/fetch/url.js
@@ -1,1 +1,0 @@
-export { URL, URLSearchParams } from "url";
