+++ packages/apollo-graphql/tsconfig.json
@@ -8,5 +8,5 @@
   },
   "include": ["src/**/*"],
   "exclude": ["**/__tests__", "**/__mocks__"],
-  "references": []
+  "references": [{ "path": "../apollo-env" }]
 }
