+++ packages/apollo-tools/tsconfig.json
@@ -5,5 +5,6 @@
     "outDir": "./lib"
   },
   "include": ["./src/**/*"],
-  "exclude": ["**/__tests__/*", "**/__mocks__/*"]
+  "exclude": ["**/__tests__/*", "**/__mocks__/*"],
+  "references": [{ "path": "../apollo-env" }]
 }
