+++ tsconfig.base.json
@@ -17,7 +17,7 @@
     "noUnusedLocals": false,
     "forceConsistentCasingInFileNames": true,
     "lib": ["es2017", "esnext.asynciterable"],
-    "types": ["node", "apollo-env"],
+    "types": ["node"],
     "baseUrl": ".",
     "paths": {
       "*": ["types/*"]
